update dependencies and remove tensorflow

+++ carla/recourse_methods/catalog/roar/library/roar.py
@@ -65,7 +65,7 @@ def roar_recourse(
     torch_model,
     x: np.ndarray,
     coeff: np.ndarray,
-    intercept: np.float,
+    intercept: float,
     cat_feature_indices: List[int],
     binary_cat_features: bool = True,
     feature_costs: Optional[List[float]] = None,
